% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentinel.R
\name{sentinel}
\alias{sentinel}
\title{Retrieve and Process Sentinel-2 Data}
\usage{
sentinel(
  pol,
  date_debut = "2018-01-01",
  date_fin = "2024-12-31",
  max_nuage = 50,
  resample = F,
  month = NULL
)
}
\arguments{
\item{pol}{An `sf` object representing the geographic region of interest.}

\item{date_debut}{Character. The start date for the query in 'YYYY-MM-DD' format. Default is '2018-01-01'.}

\item{date_fin}{Character. The end date for the query in 'YYYY-MM-DD' format. Default is '2024-12-31'.}

\item{max_nuage}{Numeric. The maximum allowed cloud cover percentage. Default is 50.}
}
\value{
A list of `SpatRaster` objects, each representing NDVI for different dates.
}
\description{
This function queries the Planetary Computer API for Sentinel-2 satellite imagery
data within a specified date range and cloud cover percentage. It processes the
imagery to calculate the Normalized Difference Vegetation Index (NDVI) for the
specified geographic region.
}
\examples{
\dontrun{
# Example usage:
library(sf)
# Define a simple polygon as the region of interest
shp <- st_as_sf(data.frame(
  geometry = st_sfc(st_polygon(list(rbind(
    c(-1, 50), c(-1, 51), c(1, 51), c(1, 50), c(-1, 50)
  )))),
  crs = 4326
))
# Retrieve and process Sentinel-2 data
ndvi_list <- sentinel(shp)
}
}
